+++ src/parser/block/extension.ts
@@ -5,17 +5,13 @@ import { placeholder } from './extension/placeholder';
 
 type SubParsers = [ExtensionParser.PlaceholderParser];
 
-const syntax = /^(~{3,})[ \t]*\n((?:[^\n]*\n)*)\1/;
-const cache = new Map<string, RegExp>();
+const syntax = /^(~{3,})[ \t]*\n(?:[^\n]*\n)*\1/;
 
 export const extension: ExtensionParser = function (source: string): Result<HTMLElement, SubParsers> {
-  const [whole, keyword, text] = source.match(syntax) || ['', '', ''];
+  const [whole] = source.match(syntax) || [''];
   if (!whole) return;
-  if (!cache.has(keyword)) {
-    void cache.set(keyword, new RegExp(`^${keyword}\s*(?:\n|$)`));
-  }
-  const [ns, rest] = compose<SubParsers, HTMLElement>([placeholder])(text) || [[], ''];
+  const [ns, rest] = compose<SubParsers, HTMLElement>([placeholder])(source) || [[], ''];
   return source.length === rest.length
     ? void 0
-    : consumeBlockEndEmptyLine<HTMLElement, SubParsers>(ns, source.slice(whole.length));
+    : consumeBlockEndEmptyLine<HTMLElement, SubParsers>(ns, rest);
 }