+++ src/parser/block/extension/placeholder.ts
@@ -1,15 +1,13 @@
 ﻿import { Result } from '../../../parser';
 import { ExtensionParser, consumeBlockEndEmptyLine } from '../../block';
+import { loop } from '../../../combinator/loop';
+import { InlineParser, inline } from '../../inline';
+import { squash } from '../../inline/text';
 
-type SubParsers = [never];
+type SubParsers = [InlineParser];
 
-export const placeholder: ExtensionParser.PlaceholderParser = function (source: string): Result<HTMLElement, SubParsers> {
+export const placeholder: ExtensionParser.PlaceholderParser = function (_: string): Result<HTMLElement, SubParsers> {
   const el = document.createElement('p');
-  while (true) {
-    const line = source.split('\n', 1)[0];
-    el.textContent += line;
-    source = source.slice(line.length + 1);
-    if (source === '') break;
-  }
-  return consumeBlockEndEmptyLine<HTMLElement, SubParsers>([el], source);
+  void el.appendChild(squash(loop(inline)("**DON'T USE `~~~` SYNTAX!!**\\\nThis extension syntax is reserved for extensibility.")![0]));
+  return consumeBlockEndEmptyLine<HTMLElement, SubParsers>([el], '');
 }